refactor: `SamplingRate`を移動
