+++ crates/voicevox_core/src/engine.rs
@@ -2,12 +2,12 @@
 
 mod acoustic_feature_extractor;
 mod audio_file;
-mod fundamental;
 mod mora_list;
+mod sampling_rate;
 pub(crate) mod talk;
 
 pub use self::audio_file::wav_from_s16le;
 pub(crate) use self::{
     acoustic_feature_extractor::PhonemeCode, audio_file::to_s16le_pcm,
-    fundamental::DEFAULT_SAMPLING_RATE,
+    sampling_rate::DEFAULT_SAMPLING_RATE,
 };
