--- crates/voicevox_core/src/engine/fundamental.rs
@@ -1,1 +1,0 @@
-pub(crate) const DEFAULT_SAMPLING_RATE: u32 = 24000;
